bugfix in betalinkr (for proportions=TRUE)
There was an error in betalinkr: betalinkr transformed the linkmx.sharedsp to proportions (summing to 1 in these subwebs) before doing the partitioning, which messed up the calculations (sometimes negative values in matrices). Results with proportions=T were thus hardly trustable.

This is now corrected and results for all methods (partitioning options, etc) should be fine also with proportions=TRUE.