betalinkr: corrected beta_S for one-mode webs
for calculating species community dissimilarity (beta_S), lower and higher marginal totals were concatenated.
This is changed now, and any names occuring both in rows and columns are combined to one entry (summing over in-going and out-going links of a species) in the community table used for calculating beta_S.
All this has no effect on interaction dissimilarity.