betalinkr: reverted one of the latest changes by BS
After some discussion with Benjamin Schwarz, I decided that the original behavior (division by zero for empty webs, and subsequent error thrown by vegan) is actually wanted behavior (a feature not a bug). In my logic, proportions of a zero-sum are not zero but NA. So to avoid the problem, one has to avoid that betalinkr is called with proportions=TRUE for empty webs, and decide oneself what the appropriate value is.