plotweb reimplementation under development, 3 functions added
I have started to reimplement the plotweb function, for simplifying future development, and created a new function for combined plots of 2 webs that are nicely aligned (aim: illustrate changes between webs).

See file "plotwebr_testingGrounds" for overview of new functions, To Do and some code for testing / development.

The functions are purposely outside the package folder, because they are not yet intended for productive use or distribution via CRAN!